Use DittoRootActor for all root actors of ditto services
