+++ services/base/src/main/java/org/eclipse/ditto/services/base/actors/DittoRootActor.java
@@ -12,16 +12,85 @@
  */
 package org.eclipse.ditto.services.base.actors;
 
+import java.io.PrintWriter;
+import java.io.StringWriter;
+import java.net.ConnectException;
+import java.util.NoSuchElementException;
+
+import org.eclipse.ditto.model.base.exceptions.DittoRuntimeException;
 import org.eclipse.ditto.services.utils.akka.logging.DittoDiagnosticLoggingAdapter;
 import org.eclipse.ditto.services.utils.akka.logging.DittoLoggerFactory;
 
 import akka.actor.AbstractActor;
+import akka.actor.ActorInitializationException;
+import akka.actor.ActorKilledException;
+import akka.actor.ActorRef;
+import akka.actor.InvalidActorNameException;
+import akka.actor.OneForOneStrategy;
 import akka.actor.Props;
+import akka.actor.SupervisorStrategy;
+import akka.japi.pf.DeciderBuilder;
 import akka.japi.pf.ReceiveBuilder;
+import akka.pattern.AskTimeoutException;
+import scala.PartialFunction;
 
+/**
+ * Abstract implementation of a root actor of a ditto service.
+ */
 public abstract class DittoRootActor extends AbstractActor {
 
     private final DittoDiagnosticLoggingAdapter log = DittoLoggerFactory.getDiagnosticLoggingAdapter(this);
+    private final PartialFunction<Throwable, SupervisorStrategy.Directive> supervisionDecider = DeciderBuilder
+            .match(NullPointerException.class, e -> {
+                log.error(e, "NullPointer in child actor: {}", e.getMessage());
+                return restartChild();
+            }).match(IllegalArgumentException.class, e -> {
+                log.warning("Illegal Argument in child actor: {}", e.getMessage());
+
+                StringWriter sw = new StringWriter();
+                PrintWriter pw = new PrintWriter(sw);
+                e.printStackTrace(pw);
+
+                log.warning("Illegal Argument in child actor: {}", sw.toString());
+                return SupervisorStrategy.resume();
+            }).match(IllegalStateException.class, e -> {
+                log.warning("Illegal State in child actor: {}", e.getMessage());
+                return SupervisorStrategy.resume();
+            }).match(IndexOutOfBoundsException.class, e -> {
+                log.warning("IndexOutOfBounds in child actor: {}", e.getMessage());
+                return SupervisorStrategy.resume();
+            }).match(NoSuchElementException.class, e -> {
+                log.warning("NoSuchElement in child actor: {}", e.getMessage());
+                return SupervisorStrategy.resume();
+            }).match(AskTimeoutException.class, e -> {
+                log.warning("AskTimeoutException in child actor: {}", e.getMessage());
+                return SupervisorStrategy.resume();
+            }).match(ConnectException.class, e -> {
+                log.warning("ConnectException in child actor: {}", e.getMessage());
+                return restartChild();
+            }).match(InvalidActorNameException.class, e -> {
+                log.warning("InvalidActorNameException in child actor: {}", e.getMessage());
+                return SupervisorStrategy.resume();
+            }).match(ActorInitializationException.class, e -> {
+                log.error(e, "ActorInitializationException in child actor: {}", e.getMessage());
+                return SupervisorStrategy.stop();
+            }).match(ActorKilledException.class, e -> {
+                log.error(e, "ActorKilledException in child actor: {}", e.message());
+                return restartChild();
+            }).match(DittoRuntimeException.class, e -> {
+                log.error(e,
+                        "DittoRuntimeException '{}' should not be escalated to ConnectivityRootActor. Simply resuming Actor.",
+                        e.getErrorCode());
+                return SupervisorStrategy.resume();
+            }).match(Throwable.class, e -> {
+                log.error(e, "Escalating above root actor!");
+                return SupervisorStrategy.escalate();
+            }).matchAny(e -> {
+                log.error("Unknown message:'{}'! Escalating above root actor!", e);
+                return SupervisorStrategy.escalate();
+            }).build();
+
+    private final SupervisorStrategy strategy = new OneForOneStrategy(true, getSupervisionDecider());
 
     @Override
     public Receive createReceive() {
@@ -30,16 +99,63 @@ public abstract class DittoRootActor extends AbstractActor {
                 .build();
     }
 
-    private void startChildActor(final StartChildActor startChildActor) {
-        log.info("Starting child actor <{}>.", startChildActor.actorName);
-        getContext().actorOf(startChildActor.props, startChildActor.actorName);
+    @Override
+    public SupervisorStrategy supervisorStrategy() {
+        return strategy;
     }
 
+    /**
+     * Gets the partial function that handles a throwable and returns a SupervisorStrategy.Directive. This PF will be
+     * used for the supervision strategy of this actor.
+     *
+     * @return the partial function.
+     */
+    protected PartialFunction<Throwable, SupervisorStrategy.Directive> getSupervisionDecider() {
+        return supervisionDecider;
+    }
+
+    /**
+     * Starts an actor in the context of this actor.
+     *
+     * @param actorName the name of the actor to start.
+     * @param props the props of the actor to start.
+     * @return the ref of the started actor.
+     */
+    protected ActorRef startChildActor(final String actorName, final Props props) {
+        log.info("Starting child actor <{}>.", actorName);
+        return getContext().actorOf(props, actorName);
+    }
+
+    /**
+     * Returns the restart supervisor strategy.
+     *
+     * @return the restart supervisor strategy.
+     */
+    protected SupervisorStrategy.Directive restartChild() {
+        log.info("Restarting child ...");
+        return SupervisorStrategy.restart();
+    }
+
+    private void startChildActor(final StartChildActor startChildActor) {
+        startChildActor(startChildActor.actorName, startChildActor.props);
+    }
+
+
+    /**
+     * Message that holds props of an actor and the name to start this actor with.
+     * Send this Object to a {@link DittoRootActor} in order to start an actor in the actor context of the root actor.
+     */
     public static class StartChildActor {
 
         private final Props props;
         private final String actorName;
 
+        /**
+         * Constructs a new instance of StartChildActor
+         *
+         * @param props the props of the child actor.
+         * @param actorName the name the child actor should be started with.
+         */
         public StartChildActor(final Props props, final String actorName) {
             this.props = props;
             this.actorName = actorName;
