+++ services/concierge/starter/src/main/java/org/eclipse/ditto/services/concierge/starter/actors/ConciergeRootActor.java
@@ -16,12 +16,10 @@ import static akka.http.javadsl.server.Directives.logRequest;
 import static akka.http.javadsl.server.Directives.logResult;
 import static org.eclipse.ditto.model.base.common.ConditionChecker.checkNotNull;
 
-import java.net.ConnectException;
 import java.time.Duration;
-import java.util.NoSuchElementException;
 import java.util.concurrent.CompletionStage;
 
-import org.eclipse.ditto.model.base.exceptions.DittoRuntimeException;
+import org.eclipse.ditto.services.base.actors.DittoRootActor;
 import org.eclipse.ditto.services.base.config.http.HttpConfig;
 import org.eclipse.ditto.services.concierge.actors.ShardRegions;
 import org.eclipse.ditto.services.concierge.actors.cleanup.CleanupStatusReporter;
@@ -43,88 +41,31 @@ import org.eclipse.ditto.services.utils.health.routes.StatusRoute;
 import org.eclipse.ditto.services.utils.persistence.mongo.MongoHealthChecker;
 
 import akka.Done;
-import akka.actor.AbstractActor;
-import akka.actor.ActorInitializationException;
-import akka.actor.ActorKilledException;
 import akka.actor.ActorRef;
-import akka.actor.ActorRefFactory;
 import akka.actor.ActorSystem;
 import akka.actor.CoordinatedShutdown;
-import akka.actor.InvalidActorNameException;
-import akka.actor.OneForOneStrategy;
 import akka.actor.Props;
 import akka.actor.Status;
-import akka.actor.SupervisorStrategy;
 import akka.cluster.Cluster;
 import akka.event.DiagnosticLoggingAdapter;
 import akka.http.javadsl.ConnectHttp;
 import akka.http.javadsl.Http;
 import akka.http.javadsl.ServerBinding;
 import akka.http.javadsl.server.Route;
-import akka.japi.pf.DeciderBuilder;
 import akka.japi.pf.ReceiveBuilder;
-import akka.pattern.AskTimeoutException;
 import akka.stream.ActorMaterializer;
 
 /**
  * The root actor of the concierge service.
  */
-public final class ConciergeRootActor extends AbstractActor {
+public final class ConciergeRootActor extends DittoRootActor {
 
     /**
      * Name of this actor.
      */
     public static final String ACTOR_NAME = "conciergeRoot";
 
-    private static final String RESTARTING_CHILD_MSG = "Restarting child...";
-
     private final DiagnosticLoggingAdapter log = LogUtil.obtain(this);
-    private final SupervisorStrategy supervisorStrategy = new OneForOneStrategy(true, DeciderBuilder
-            .match(NullPointerException.class, e -> {
-                log.error(e, "NullPointer in child actor: {}", e.getMessage());
-                log.info(RESTARTING_CHILD_MSG);
-                return SupervisorStrategy.restart();
-            }).match(IllegalArgumentException.class, e -> {
-                log.warning("Illegal Argument in child actor: {}", e.getMessage());
-                return SupervisorStrategy.resume();
-            }).match(IndexOutOfBoundsException.class, e -> {
-                log.warning("IndexOutOfBounds in child actor: {}", e.getMessage());
-                return SupervisorStrategy.resume();
-            }).match(IllegalStateException.class, e -> {
-                log.warning("Illegal State in child actor: {}", e.getMessage());
-                return SupervisorStrategy.resume();
-            }).match(NoSuchElementException.class, e -> {
-                log.warning("NoSuchElement in child actor: {}", e.getMessage());
-                return SupervisorStrategy.resume();
-            }).match(AskTimeoutException.class, e -> {
-                log.warning("AskTimeoutException in child actor: {}", e.getMessage());
-                return SupervisorStrategy.resume();
-            }).match(ConnectException.class, e -> {
-                log.warning("ConnectException in child actor: {}", e.getMessage());
-                log.info(RESTARTING_CHILD_MSG);
-                return SupervisorStrategy.restart();
-            }).match(InvalidActorNameException.class, e -> {
-                log.warning("InvalidActorNameException in child actor: {}", e.getMessage());
-                return SupervisorStrategy.resume();
-            }).match(ActorInitializationException.class, e -> {
-                log.error(e, "ActorInitializationException in child actor: {}", e.getMessage());
-                return SupervisorStrategy.stop();
-            }).match(ActorKilledException.class, e -> {
-                log.error(e, "ActorKilledException in child actor: {}", e.message());
-                log.info(RESTARTING_CHILD_MSG);
-                return SupervisorStrategy.restart();
-            }).match(DittoRuntimeException.class, e -> {
-                log.error(e,
-                        "DittoRuntimeException '{}' should not be escalated to RootActor. Simply resuming Actor.",
-                        e.getErrorCode());
-                return SupervisorStrategy.resume();
-            }).match(Throwable.class, e -> {
-                log.error(e, "Escalating above root actor!");
-                return SupervisorStrategy.escalate();
-            }).matchAny(e -> {
-                log.error("Unknown message:'{}'! Escalating above root actor!", e);
-                return SupervisorStrategy.escalate();
-            }).build());
 
     @SuppressWarnings("unused")
     private <C extends ConciergeConfig> ConciergeRootActor(final C conciergeConfig,
@@ -142,12 +83,12 @@ public final class ConciergeRootActor extends AbstractActor {
         final ActorRef conciergeForwarder = context.findChild(ConciergeForwarderActor.ACTOR_NAME).orElseThrow(() ->
                 new IllegalStateException("ConciergeForwarder could not be found"));
 
-        final ActorRef cleanupCoordinator = startClusterSingletonActor(context,
+        final ActorRef cleanupCoordinator = startClusterSingletonActor(
                 EventSnapshotCleanupCoordinator.ACTOR_NAME,
                 EventSnapshotCleanupCoordinator.props(conciergeConfig.getPersistenceCleanupConfig(), pubSubMediator,
                         shardRegions));
 
-        final ActorRef healthCheckingActor = startHealthCheckingActor(context, conciergeConfig, cleanupCoordinator);
+        final ActorRef healthCheckingActor = startHealthCheckingActor(conciergeConfig, cleanupCoordinator);
 
         bindHttpStatusRoute(healthCheckingActor, conciergeConfig.getHttpConfig(), materializer);
     }
@@ -177,14 +118,13 @@ public final class ConciergeRootActor extends AbstractActor {
     }
 
 
-    private static ActorRef startClusterSingletonActor(final ActorContext context, final String actorName,
-            final Props props) {
+    private ActorRef startClusterSingletonActor(final String actorName, final Props props) {
 
-        return ClusterUtil.startSingleton(context, ConciergeMessagingConstants.CLUSTER_ROLE, actorName, props);
+        return ClusterUtil.startSingleton(getContext(), ConciergeMessagingConstants.CLUSTER_ROLE, actorName, props);
     }
 
-    private static ActorRef startHealthCheckingActor(final ActorContext context,
-            final ConciergeConfig conciergeConfig, final ActorRef cleanupCoordinator) {
+    private ActorRef startHealthCheckingActor(final ConciergeConfig conciergeConfig,
+            final ActorRef cleanupCoordinator) {
 
         final HealthCheckConfig healthCheckConfig = conciergeConfig.getHealthCheckConfig();
 
@@ -197,18 +137,13 @@ public final class ConciergeRootActor extends AbstractActor {
         }
         final HealthCheckingActorOptions healthCheckingActorOptions = hcBuilder.build();
 
-        final ActorRef cleanupCoordinatorProxy = ClusterUtil.startSingletonProxy(context,
+        final ActorRef cleanupCoordinatorProxy = ClusterUtil.startSingletonProxy(getContext(),
                 ConciergeMessagingConstants.CLUSTER_ROLE, cleanupCoordinator);
 
-        return startChildActor(context, DefaultHealthCheckingActorFactory.ACTOR_NAME,
+        return startChildActor(DefaultHealthCheckingActorFactory.ACTOR_NAME,
                 DefaultHealthCheckingActorFactory.props(healthCheckingActorOptions,
                         MongoHealthChecker.props(),
                         CleanupStatusReporter.props(cleanupCoordinatorProxy)));
-    }
-
-    private static ActorRef startChildActor(final ActorRefFactory context, final String actorName, final Props props) {
-
-        return context.actorOf(props, actorName);
     }
 
     private static Route createRoute(final ActorSystem actorSystem, final ActorRef healthCheckingActor) {
@@ -220,18 +155,14 @@ public final class ConciergeRootActor extends AbstractActor {
     }
 
     @Override
-    public SupervisorStrategy supervisorStrategy() {
-        return supervisorStrategy;
-    }
-
-    @Override
     public Receive createReceive() {
-        return ReceiveBuilder.create()
-                .match(Status.Failure.class, f -> log.error(f.cause(), "Got failure <{}>!", f))
-                .matchAny(m -> {
-                    log.warning("Unknown message <{}>.", m);
-                    unhandled(m);
-                }).build();
+        return super.createReceive()
+                .orElse(ReceiveBuilder.create()
+                        .match(Status.Failure.class, f -> log.error(f.cause(), "Got failure <{}>!", f))
+                        .matchAny(m -> {
+                            log.warning("Unknown message <{}>.", m);
+                            unhandled(m);
+                        }).build());
     }
 
     private void bindHttpStatusRoute(final ActorRef healthCheckingActor, final HttpConfig httpConfig,
