+++ services/gateway/starter/src/main/java/org/eclipse/ditto/services/gateway/starter/GatewayRootActor.java
@@ -12,15 +12,13 @@
  */
 package org.eclipse.ditto.services.gateway.starter;
 
-import java.net.ConnectException;
 import java.time.Duration;
-import java.util.NoSuchElementException;
 import java.util.concurrent.CompletionStage;
 import java.util.function.Supplier;
 
-import org.eclipse.ditto.model.base.exceptions.DittoRuntimeException;
 import org.eclipse.ditto.model.base.headers.DittoHeadersSizeChecker;
 import org.eclipse.ditto.protocoladapter.HeaderTranslator;
+import org.eclipse.ditto.services.base.actors.DittoRootActor;
 import org.eclipse.ditto.services.base.config.limits.LimitsConfig;
 import org.eclipse.ditto.services.gateway.endpoints.config.HttpConfig;
 import org.eclipse.ditto.services.gateway.endpoints.directives.auth.DittoGatewayAuthenticationDirectiveFactory;
@@ -29,10 +27,10 @@ import org.eclipse.ditto.services.gateway.endpoints.routes.RootRoute;
 import org.eclipse.ditto.services.gateway.endpoints.routes.devops.DevOpsRoute;
 import org.eclipse.ditto.services.gateway.endpoints.routes.health.CachingHealthRoute;
 import org.eclipse.ditto.services.gateway.endpoints.routes.policies.PoliciesRoute;
-import org.eclipse.ditto.services.gateway.endpoints.routes.things.ThingsSseRouteBuilder;
 import org.eclipse.ditto.services.gateway.endpoints.routes.stats.StatsRoute;
 import org.eclipse.ditto.services.gateway.endpoints.routes.status.OverallStatusRoute;
 import org.eclipse.ditto.services.gateway.endpoints.routes.things.ThingsRoute;
+import org.eclipse.ditto.services.gateway.endpoints.routes.things.ThingsSseRouteBuilder;
 import org.eclipse.ditto.services.gateway.endpoints.routes.thingsearch.ThingSearchRoute;
 import org.eclipse.ditto.services.gateway.endpoints.routes.websocket.WebSocketRoute;
 import org.eclipse.ditto.services.gateway.health.DittoStatusAndHealthProviderFactory;
@@ -65,16 +63,11 @@ import org.eclipse.ditto.services.utils.health.routes.StatusRoute;
 import org.eclipse.ditto.services.utils.protocol.ProtocolAdapterProvider;
 
 import akka.Done;
-import akka.actor.AbstractActor;
-import akka.actor.ActorKilledException;
 import akka.actor.ActorRef;
 import akka.actor.ActorSystem;
 import akka.actor.CoordinatedShutdown;
-import akka.actor.InvalidActorNameException;
-import akka.actor.OneForOneStrategy;
 import akka.actor.Props;
 import akka.actor.Status;
-import akka.actor.SupervisorStrategy;
 import akka.cluster.Cluster;
 import akka.dispatch.MessageDispatcher;
 import akka.event.DiagnosticLoggingAdapter;
@@ -84,15 +77,13 @@ import akka.http.javadsl.Http;
 import akka.http.javadsl.ServerBinding;
 import akka.http.javadsl.server.Directives;
 import akka.http.javadsl.server.Route;
-import akka.japi.pf.DeciderBuilder;
 import akka.japi.pf.ReceiveBuilder;
-import akka.pattern.AskTimeoutException;
 import akka.stream.ActorMaterializer;
 
 /**
  * The Root Actor of the API Gateway's Akka ActorSystem.
  */
-final class GatewayRootActor extends AbstractActor {
+final class GatewayRootActor extends DittoRootActor {
 
     /**
      * The name of this Actor in the ActorSystem.
@@ -101,53 +92,7 @@ final class GatewayRootActor extends AbstractActor {
 
     private static final String AUTHENTICATION_DISPATCHER_NAME = "authentication-dispatcher";
 
-    private static final String CHILD_RESTART_INFO_MSG = "Restarting child ...";
-
     private final DiagnosticLoggingAdapter log = LogUtil.obtain(this);
-
-    private final SupervisorStrategy strategy = new OneForOneStrategy(true, DeciderBuilder
-            .match(NullPointerException.class, e -> {
-                log.error(e, "NullPointer in child actor: {}", e.getMessage());
-                log.info(CHILD_RESTART_INFO_MSG);
-                return SupervisorStrategy.restart();
-            }).match(IllegalArgumentException.class, e -> {
-                log.warning("Illegal Argument in child actor: {}", e.getMessage());
-                return SupervisorStrategy.resume();
-            }).match(IndexOutOfBoundsException.class, e -> {
-                log.warning("IndexOutOfBounds in child actor: {}", e.getMessage());
-                return SupervisorStrategy.resume();
-            }).match(IllegalStateException.class, e -> {
-                log.warning("Illegal State in child actor: {}", e.getMessage());
-                return SupervisorStrategy.resume();
-            }).match(NoSuchElementException.class, e -> {
-                log.warning("NoSuchElement in child actor: {}", e.getMessage());
-                return SupervisorStrategy.resume();
-            }).match(AskTimeoutException.class, e -> {
-                log.warning("AskTimeoutException in child actor: {}", e.getMessage());
-                return SupervisorStrategy.resume();
-            }).match(ConnectException.class, e -> {
-                log.warning("ConnectException in child actor: {}", e.getMessage());
-                log.info(CHILD_RESTART_INFO_MSG);
-                return SupervisorStrategy.restart();
-            }).match(InvalidActorNameException.class, e -> {
-                log.warning("InvalidActorNameException in child actor: {}", e.getMessage());
-                return SupervisorStrategy.resume();
-            }).match(DittoRuntimeException.class, e -> {
-                log.error(e,
-                        "DittoRuntimeException <{}> should not be escalated to GatewayRootActor. Simply resuming Actor.",
-                        e.getErrorCode());
-                return SupervisorStrategy.resume();
-            }).match(ActorKilledException.class, e -> {
-                log.error(e, "ActorKilledException in child actor: {}", e.message());
-                log.info(CHILD_RESTART_INFO_MSG);
-                return SupervisorStrategy.restart();
-            }).match(Throwable.class, e -> {
-                log.error(e, "Escalating above root actor!");
-                return SupervisorStrategy.escalate();
-            }).matchAny(e -> {
-                log.error("Unknown message: '{}'! Escalating above root actor!", e);
-                return SupervisorStrategy.escalate();
-            }).build());
 
     private final CompletionStage<ServerBinding> httpBinding;
 
@@ -241,28 +186,19 @@ final class GatewayRootActor extends AbstractActor {
     }
 
     @Override
-    public SupervisorStrategy supervisorStrategy() {
-        return strategy;
-    }
-
-    @Override
     public Receive createReceive() {
-        return ReceiveBuilder.create()
-                .match(Status.Failure.class, f -> log.error(f.cause(), "Got failure: {}", f))
-                .matchEquals(GatewayHttpReadinessCheck.READINESS_ASK_MESSAGE, msg -> {
-                    final ActorRef sender = getSender();
-                    httpBinding.thenAccept(binding -> sender.tell(
-                            GatewayHttpReadinessCheck.READINESS_ASK_MESSAGE_RESPONSE, ActorRef.noSender()));
-                })
-                .matchAny(m -> {
-                    log.warning("Unknown message: {}", m);
-                    unhandled(m);
-                }).build();
-    }
-
-    private ActorRef startChildActor(final String actorName, final Props props) {
-        log.info("Starting child actor <{}>.", actorName);
-        return getContext().actorOf(props, actorName);
+        return super.createReceive()
+                .orElse(ReceiveBuilder.create()
+                        .match(Status.Failure.class, f -> log.error(f.cause(), "Got failure: {}", f))
+                        .matchEquals(GatewayHttpReadinessCheck.READINESS_ASK_MESSAGE, msg -> {
+                            final ActorRef sender = getSender();
+                            httpBinding.thenAccept(binding -> sender.tell(
+                                    GatewayHttpReadinessCheck.READINESS_ASK_MESSAGE_RESPONSE, ActorRef.noSender()));
+                        })
+                        .matchAny(m -> {
+                            log.warning("Unknown message: {}", m);
+                            unhandled(m);
+                        }).build());
     }
 
     private static Route createRoute(final ActorSystem actorSystem,
